jichcha qoldi tugashiga

+++ controllers/likes.controller.js
@@ -29,8 +29,8 @@ async function findAll(req, res) {
     loggerInfo.info(`Method: ${req.method};  Saccessfully FindAll Like;`);
     res.status(200).send({ data: all });
   } catch (e) {
-    loggerError.error(`ERROR: ${e};  Method: ${req.method};  Likes-FindOne`);
-    res.status(500).json({ error: e });
+    loggerError.error(`ERROR: ${e.message };  Method: ${req.method};  Likes-FindOne`);
+    res.status(500).json({ error: e.message });
   }
 }
 
@@ -61,20 +61,21 @@ async function findOne(req, res) {
     loggerInfo.info(`Method: ${req.method};  Saccessfully FindOne Like;`);
     res.status(200).json({ data: one });
   } catch (e) {
-    loggerError.error(`ERROR: ${e};  Method: ${req.method};  Likes-FindOne`);
-    res.status(500).json({ error: e });
+    loggerError.error(`ERROR: ${e.message};  Method: ${req.method};  Likes-FindOne`);
+    res.status(500).json({ error: e.message });
   }
 }
 
 async function create(req, res) {
   try {
+    req.body.userId = req.user.id
     let { error, value } = LikeValidation.validate(req.body);
 
     if (error) {
       loggerError.error(
         `ERROR: ${error.details[0].message};  Method: ${req.method};  Like-Create`
       );
-      return res.json({ error: error.details[0].message });
+      return res.status(400).json({ error: error.details[0].message });
     }
 
     let checkUser = await User.findOne({where: {id: value.userId}})
@@ -109,22 +110,24 @@ async function create(req, res) {
       ],
     });
 
-
-
     if (check) {
       loggerError.error(
         `ERROR: Siz Like Bosib Bo'lgansiz😊;  Method: ${req.method};  Like-Create`
       );
-      return res.status(409).json({ error: "Siz Like Bosib Bo'lgansiz😊" });
+      return res.status(409).json({ error: "You have already liked" });
     }
 
-    await Like.create(value);
+    if (req.user.id !== value.userId) {
+      return res.status(403).json({error: "You can't like someone else's name."})
+    }
 
-    loggerInfo.info(`Method: ${req.method};  Saccessfully Create Like;`);
+    await Like.create(value)
+
+    loggerInfo.info(`Method: ${req.method};  Successfully Create Like;`);
     res.status(201).json({ message: "Liked  Successfully" });
   } catch (e) {
-    loggerError.error(`ERROR: ${e};  Method: ${req.method};  Likes-Create`);
-    res.status(500).json({ error: e });
+    loggerError.error(`ERROR: ${e.message};  Method: ${req.method};  Likes-Create`);
+    res.status(500).json({ error: e.message });
   }
 }
 
@@ -137,16 +140,20 @@ async function remove(req, res) {
       loggerError.error(
         `ERROR: Liked Not Found;  Method: ${req.method};  Like-Delete`
       );
-      return res.status(404).json({ error: "Liked Not Found" });
+      return res.status(404).json({ error: "Like Not Found" });
+    }
+
+    if (req.user.id !== check.userId) {
+      return res.status(403).json({error: "You can't delete someone else's like."})
     }
 
     await Like.destroy({ where: { id } });
 
-    loggerInfo.info(`Method: ${req.method};  Saccessfully Delete Like;`);
-    res.status(200).json({ message: "Delete Like Successfully" });
+    loggerInfo.info(`Method: ${req.method};  Successfully Deleted Like;`);
+    res.status(200).json({ message: "Deleted Like Successfully" });
   } catch (e) {
-    loggerError.error(`ERROR: ${e};  Method: ${req.method};  Likes-Delete`);
-    res.status(500).json({ error: e });
+    loggerError.error(`ERROR: ${e.message};  Method: ${req.method};  Likes-Delete`);
+    res.status(500).json({ error: e.message });
   }
 }
 
@@ -170,10 +177,10 @@ async function Search(req, res) {
       order: [["createdAt", create.toUpperCase()]],
     });
 
-    loggerInfo.info(`Method: ${req.method};  Saccessfully Search Like;`);
+    loggerInfo.info(`Method: ${req.method};  Su ccessfully Searched Like;`);
     return res.status(200).json({ data: check });
   } catch (e) {
-    loggerError.error(`ERROR: ${e};  Method: ${req.method};  Likes-Search`);
+    loggerError.error(`ERROR: ${e.message};  Method: ${req.method};  Likes-Search`);
     res.status(500).send({ error: e.message });
   }
 }
